Separate create and start function for provider
This PR make sure container creation happen before the start.
`podman run` command creates and start the container with single
command so in case the container is stopped run command it going
to throw error of existing container name. This pr makes easy to
add logic for starting a container if exist.

Signed-off-by: Praveen Kumar <kumarpraveen.nitdgp@gmail.com>