Add volume option for docker with read-write permissions
The location /var/lib/containers/storage is only relevant for podman to share overlays between host
and container. Docker uses overlay2 and that's why overlay sharing won't work.

By specifying /var/lib/microshift for docker with the read-write option the /host-container volume mount is available
and can be used for the lock files of the overlay-images and overlay-layers. The /var/lib/microshift directory is created
automaticly.

Signed-off-by: Harm Jan Stam <harmjan+git@harmjan.com>